requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
insightface==0.7.3
qdrant-client==1.7.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
onnxruntime==1.16.3